Add temacs make rule dep on libremacs.a
This ensures temacs is relinked if a new libremacs.a is present, ie if
rust-src has changed.

In my original patch I somehow left this out... Anyway, this makes sure
temacs is relinked if the cargo is rerun with different flags OR if the
libremacs.a archive file is updated.

Previously it only depended on cargo having been rerun with different
flags without changing the libremacs.a archive that temacs had been
built with. ie I handled the edge case and forgot about the rest...