Fix issues on 32 bits platforms.
This is a big change to the remacs Rust codebase, it basically changes
everything.

The 'defun!' macro was modified to automatically create a wrapper
function that will be used outside of C, it means that you don't need to
actually use 'unsafe' code in your functions, however the existing
functions need to be updated to work without 'unsafe' blocks.

Also a new macro was added, its called 'defun_many!', it is similar to
the normal 'defun 'macro however this only works for functions that
accept a dynamic list of parameters. It nicely creates a wrapper
function like the 'defun!' to pass a slice of arguments (LispObject) to
the actual function.

I found that the code in math.rs is a bit complicated, specially
the arith_driver function, it needs to be refactored, i don't know if
that function actually works correctly.

The src/Makefile.in file was also modified to rebuild if the files
inside remacs-sys changed, including Cargo.toml.