Refactor frame handling (#584)
Break the core logic of `framep` out into a separate function. Use
this function internally since we already have `LispFrameRef` objects.

This allows `frame_live_p` to not duplicate the frame checks again in
`framep`.

`window-system` should just return a `LispObject`. There is no reason
to convert to a Symbol and immediately turn it back into a
`LispObject`.