Fix obarray checks in intern and mapatoms
We rely on check_obarray to verify that a value is an obarray. That
function assumed the argument was a vector (it used an unchecked
functions inappropriately).

This caused intern and mapatoms to crash when called with non-obarray
arguments.

Also fix an issue with intern where it assumed a non-vector means it
should use the global obarray, whereas it should only the global
obarray when called with nil.

Add tests for all these cases.

Fixes #731.