use bindgen to auto-generate 99% of the bindinges
This patch touches a lot of stuff, because the bindings are very
exact. Pointers are mutable by default, unions that we were just
ignoring are now there (see Lisp_Float and similar things), some
constants became enums, some enums became constants, some things have
worse names (pvec_type, for instance), etc. Many things are much
better, such as bitfields and the vectors and vectorlike objects.

The only major downside to this patch is the compile times. It adds 8
minutes to the build time (on my machine, anyway). Four minutes while
building libremacs to generate the bindings, then another four minutes
to generate the bindings again when building the Rust tests. It's a
separate binary, you see, so it has a separate output directory and
everything. Fun times. I'll look for a way to improve that.