Bump bindgen version
This includes the fix
in https://github.com/rust-lang-nursery/rust-bindgen/pull/1419
which means the rustified_enum calls introduced by
https://github.com/Wilfred/remacs/pull/915
are no longer necessary --- and in general things should be
more future-proof now; if some remacs dependency introduces a new enum,
it won't need to be blacklisted to prevent faulty bindgen-generated
code from breaking things.

The equivalent of bindgen's `.blacklist_type` (which used to not actually care
that the blacklisted thing was a type, despite its name)
is now `.blacklist_item`.