Refresh usage of LispCons iterators. (#1127)
This builds on #1118.

This PR is the result of examining each use of the iterator compared to the original Emacs source.

What we have now is:
 - Does the iteration check if the last item to be inspected is a valid CONS? This is `LispConsEndChecks::on`
 - Does the iteration check for a circular list, IOW one that loops back on itself? This is `LispConsCircularChecks`. The value of `on` means the check happens and if a loop is found a signal is raised. `safe` means the check is done but all it does is exit. `off` means no checking at all.

In the Rust code we used to mix up ends checks with circular checks. Many of the places where we used `iter_tails_safe` what we really wanted was either completely unchecked or just end checking. Only occasionally did we want Ends off and Circular safe.

Now every call to get an iterator is explicit in what it needs. This is more verbose. However, the matrix of potential functions is now 6 methods -- circular safe and ends checks no, circular safe and ends checks yes, .....

These three are the most common pairs:
EndChecks on and Circular Checks on -- the original iter_tails
EndsChecks off and Circular Checks safe -- iter_tails_safe
EndsChecks off and Circular Checks off -- iter_tails_unchecked

Porting guidelines
-----------------

If the C code contains:
 * `FOR_EACH_TAIL` use `LispConsCircularChecks::on`
 * `FOR_EACH_TAIL_SAFE` means `LispConsCircularChecks::safe`
 * manual for / while loops are _usually_ `LispConsCircularChecks::off` but you need to look to see if there are references to 'tortoise'. If there are, then circular checks are being used and `LispConsCircularChecks::on` should be used.

If the loop ends with `CHECK_LIST_END` or a check that the last item inspected is `CONSP` use `LispConsEndChecks::on` otherwise use off.

The choice of whether to use `iter_tails` or `iter_cars` depends on what the code is doing. The C code does not have a `cars iteration` concept. So using `iter_tails` is a good place to start. If the loop only really uses the `car` of the CONS cell then `iter_cars` is better.

Often, we end up needing to look at the remaining tail to see if the list exited early. This is possible by calling at the iterator's `rest()` method. However, if you used the iterator in a for loop the iterator was moved. You need to save the iterator before the loop and use `it.next()` instead of `for item in it`.