Correct base64 multibyte tests
It looks like there was some confusion when originally porting the
code. When we pass `multibyte` as `true` to the base64 encode and
decode functions, we indicate that the input and output respectively
follow emacs' own encoding for multibyte characters.

The tests we had treated a UTF-8 string as though it were multibyte
while in emacs parlance it would be treated as unibyte because we want
these functions to treat it as a raw buffer.

Correct these tests so the test string matches the base64 data and we
correctly test against unibyte vs multibyte handling.