Fix LispMarkerRef tagging when converting to LispObject and make some use of it (#1177)
The implmentation of `From<LispMakerRef> for LispObject` transmutes the memory which means we fail to tag the pointer. As the receiver expects it to be tagged it thinks it's a `Lisp_Symbol` which confuses the code quite a bit and can cause assertions to fail.

Fix that and tag the pointer as "misc". Since this is fixed, make use of the conversions by accepting and returning `LispMarkerRef` in some lisp function implementations which would crash `temacs` before this change.

---

This tagging error is the cause of #1167 and it's probably related to  #820 and #1145 though I can't be positive as those are assertions on the contents.