Break up implementation of Debug trait. (#1158)
To facilitate this, introduce `force_foo` methods which are analogous to `XFOO` in the C code. These can replace some of our usage of `as_foo_or_error()` when we are certain of the type of the object.

What I'd like to see is a bit of bikeshedding over the naming I used. is `force_foo` clear enough? I wanted the name to be short so it did not use up too much space in code and detract from the real work. The C `XFOO` is fairly perfect in that regard.