Make rustize-emacs function comment regex reluctant
Before: The script crashes when run on `set_window_hscroll` in
`src/window.c`, because the `<comment>` group consumes both the comment
before the function definition and the two comments inside the function
body.

After: The `<comment>` group has a reluctant modifier so it consumes as
few characters as possible.