Port font-get-system-font and font-get-system-normal-font (#1467)
* Port font-get-system-font and font-get-system-normal-font

* Modify build.rs to handle ignored files

xsettings.rs shouldn't be compiled at all in builds without the
"window-system-x11" feature. Adding a #[cfg] guard to its module
declaration in lib.rs works to remove it from compilation. However,
it doesn't prevent build.rs from generating bindings for it, which leads
to compilation errors.

These changes make it so build.rs maintains a list of ignored files
that varies with build features. This prevents binding generation for
ignored files, but it does mean we need to change build.rs to include
any new files that should be ignored by Cargo depending on which
features we're using.