Make Codepoint a newtype (#1502)
Change Codepoint from a type alias for u32 to a newtype. This enables
adding methods to the type, which several existing functions have been
converted to. It also enables conversion to and from lisp objects using
the From trait, which means the type can be used in lisp function
signatures. Types have been changed in lisp functions where appropriate.
 
 Additional changes:
 * Make LispBufferRef::fetch_{,multibyte_}char return Codepoint instead
 of c_int
 * Remove LispObject::as_character_or_error