lisp: add a cast() method to ExternalPtr<T>
This casts one `ExternalPtr<T>` to a `ExternalPtr<U>` and lets the
compiler take care of discovering the types.

This avoids having to convert into a raw type and back and avoid
having to type out the type when these conversions happen inside
extremely simple methods to check and convert.